Error fixes for 17th solutions
Fixed all build errors that wass skiped during last 2 commits.
