+++ PipesProvider/PipesProvider.csproj
@@ -42,18 +42,22 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ClientAPI.cs" />
+    <Compile Include="Client\ClientAPI.cs" />
+    <Compile Include="Handlers\DNS.cs" />
+    <Compile Include="Handlers\Query.cs" />
+    <Compile Include="Handlers\Service.cs" />
     <Compile Include="Networking\Info.cs" />
     <Compile Include="Networking\RoutingInstruction.cs" />
+    <Compile Include="Security\Crypto.cs" />
     <Compile Include="Security\General.cs" />
     <Compile Include="Security\LogonConfig.cs" />
     <Compile Include="Security\LSA.cs" />
     <Compile Include="Security\SecurityLevel.cs" />
-    <Compile Include="Networking\TransmissionLine.cs" />
+    <Compile Include="Client\TransmissionLine.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="QueryContainer.cs" />
-    <Compile Include="ServerAPI.cs" />
-    <Compile Include="Networking\ServerTransmissionMeta.cs" />
+    <Compile Include="Server\QueryContainer.cs" />
+    <Compile Include="Server\ServerAPI.cs" />
+    <Compile Include="Server\ServerTransmissionController.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\UniformQueries\UniformQueries.csproj">
@@ -61,5 +65,6 @@
       <Name>UniformQueries</Name>
     </ProjectReference>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
