+++ TestClient/Client.cs
@@ -17,6 +17,7 @@ using System.Threading;
 using Microsoft.Win32.SafeHandles;
 using PipesProvider.Networking;
 using PipesProvider.Security;
+using PipesProvider.Client;
 
 namespace TestClient
 {
@@ -99,7 +100,7 @@ namespace TestClient
             #endregion
 
             // Close all active lines. Without this operation thread will be hanged.
-            PipesProvider.ClientAPI.CloseAllTransmissionLines();
+            ClientAPI.CloseAllTransmissionLines();
 
             // Whait until close.
             Console.WriteLine("Press any key to exit...");
