+++ UniformServer/Queries/GET_PUBLIC_KEY.cs
@@ -40,11 +40,8 @@ namespace UniformServer.Queries
 
         public void Execute(QueryPart[] queryParts)
         {
-            // Look for token.
-            API.TryGetParamValue("token", out QueryPart token, queryParts);
-
             // Create public key as answer.
-            string answer = SecurityAPI.PublicKeyXML;
+            string answer = PipesProvider.Security.Crypto.PublicKeyXML;
 
             // Open answer chanel on server and send message.
             BaseServer.SendAnswer(answer, queryParts);
